Add files via upload
A Sales Dashboard Excel enables businesses to visualise their real-time Sales data and enhance their productivity. In this project, I utilised the various critical Sales KPIs that I kept in mind when configuring a Sales Dashboard Excel based on my use case. I have utilised a variety of Excel functions and Pivot tables for my project. In addition, I have utilised the Vlookup, offset, and Rank functions for data cleansing and analysis. Throughout this effort, I've learned a great deal of new vocabulary and functions, and I've encountered a great deal of difficulty, for which I've turned to YouTube and Google for assistance. It took me four hours to complete this assignment because it was my first Excel project, which requires a great deal of time. By selecting months and years from slicer, it is possible to view the profit, sales, and overall profit, among other things. THankyou